pyproject.toml
[project]
name = "tianzhu-report"
version = "0.1.0"
description = "天柱专项报表系统 - 基于标准 src 布局"
readme = "README.md"
requires-python = ">=3.11"
# 核心运行依赖：这些是代码中直接 import 的顶级包
dependencies = [
    "pandas>=2.3.1",
    "numpy>=2.3.2",
    "streamlit>=1.49.0",
    "plotly",                # 绘图核心，requirements 中缺失但代码有引用
    "sqlalchemy>=2.0.42",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "openpyxl",              # Excel 读取引擎
    "comtypes",              # 用于 Excel COM 方案
    "pyarrow>=21.0.0",       # Parquet 快照支持
    "altair>=5.5.0",         # Streamlit 辅助绘图
    "pymupdf>=1.26.7",
    "pydantic-settings>=2.12.0",
]

[tool.uv]
# 开发依赖：仅在本地开发和测试时需要
dev-dependencies = [
    "pytest",
    "ipykernel",
    "jupyterlab",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vivo_project"]
